requirements.txt
pymobiledevice3>=4.0.0
appium-python-client>=4.0.0
mcp>=0.1.0
pydantic>=2.6.0
loguru>=0.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pillow>=10.0.0
